% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc4.R
\name{dgc4}
\alias{dgc4}
\title{Univariate 4th Order Gram-Charlier Density}
\usage{
dgc4(x, cmom, rm.neg = TRUE, log = FALSE, debug = FALSE)
}
\arguments{
\item{x}{vector of density quantiles.}

\item{cmom}{vector of first 4 central moments of distribution}

\item{rm.neg}{logical; if TRUE clips the density at smallest calculated non-negative value.}

\item{log}{logical; if TRUE returns approximating density on log scale.}
}
\value{
Vector of pdf or log(pdf).
}
\description{
Univariate 4th Order Gram-Charlier Density
}

