% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussCop-package.R
\docType{package}
\name{GaussCop}
\alias{GaussCop}
\alias{GaussCop-package}
\title{Fitting, Evaluating, and Sampling from the Gaussian Copula Distribution.}
\description{
Fitting, Evaluating, and Sampling from the Gaussian Copula Distribution.
}

