% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcop.R
\name{gcop}
\alias{dgcop}
\alias{gcop}
\alias{rgcop}
\title{The Gaussian Copula Distribution.}
\usage{
dgcop(X, gCop, log = FALSE, decomp = FALSE)

rgcop(n, gCop)
}
\arguments{
\item{X}{\code{n x p} values at which to evaluate the \code{p}-dimensional density.}

\item{gCop}{the \code{gcop} model specification.}

\item{decomp}{logical; if TRUE returns the normalized residuals \code{Z}, their log-density \code{zlpdf}, and the log-jacobian \code{zljac}, such that the total log-density is \code{xldens = zldens + zljac}.  See details.}

\item{n}{number of random samples to draw.}
}
\value{
A vector of densities or of random samples.
}
\description{
The Gaussian Copula Distribution.
}
\details{
The density of Gaussian Copula distribution is
\deqn{\code{dgcop(X) = MVN(Z | Rho) * prod(f_i(X_i)) / prod(dnorm(Z_i))},}
where MVN is the PDF of a multivariate normal with mean 0 and correlation \code{Rho}, \code{f_i(x_i)} are the marginal PDFs of the original data, \code{dnorm} is teh PDF of a standard normal, and \code{Z_i = qnorm(F_i(X_i))} is the normal quantile corresponding to \code{X_i}.
}

