% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcopSub.R
\name{gcopSub}
\alias{gcopSub}
\title{Marginal subset of a gcop object.}
\usage{
gcopSub(gCop, subset)
}
\arguments{
\item{gCop}{A \code{gcop} object.}

\item{subset}{Vector specifying which subset of variables to keep.  Default is all.}
}
\value{
A \code{gcop} with variables possibly removed.
}
\description{
Marginal subset of a gcop object.
}
