% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelXD.R
\name{kernelXD}
\alias{kernelXD}
\title{Construct an \code{xDensity} representation of a kernel smoothing estimator.}
\usage{
kernelXD(x, n = 512, from, to, mean, sd, any0 = FALSE, ...)
}
\arguments{
\item{x}{Vector of samples from the underlying distribution.}

\item{n, from, to}{Optional arguments to \code{\link[stats]{density}} which are used to set the grid on which to evaluate the kernel estimator.}

\item{mean, sd}{Optional mean and standard deviation arguments for the extended density.  Default to the mean and standard deviation of \code{x}.}

\item{any0}{Logical; if \code{FALSE} forces the support of the density to be the real line.}

\item{...}{Additional arguments to \code{\link[stats]{density}}.}
}
\value{
An \code{xDensity} object.
}
\description{
Construct an \code{xDensity} representation of a kernel smoothing estimator.
}
\examples{
# xDensity approximation to a noncentral-t distribution

# true parameters
lambda <- rnorm(1) # noncentrality parameter
nu <- runif(1, 4, 6) # degrees of freedom

# simulate data
x <- rt(1e4, df = nu, ncp = lambda)

# xDensity approximation
xDens <- kernelXD(x)

# true vs approximate PDF
curve(dt(x, df = nu, ncp = lambda),
      from = min(x), to = max(x), ylab = "Density")
curve(dXD(x, xDens = xDens), add = TRUE, col = "red")
legend("topleft", legend = c("True PDF", "xDensity Approx."),
       fill = c("black", "red"))
}
