% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdensity.R
\name{matrixXD}
\alias{matrixXD}
\title{Extend a matrix representation of a density to an \code{xDens} object.}
\usage{
matrixXD(XY, mean, sd)
}
\arguments{
\item{XY}{2-column matrix representation of a univariate density.  The first column is a grid of x values giving the centers of bins.   The second column are the corresponding densities.}

\item{mean, sd}{Optional mean and standard deviation of the Normal distribution to use outside the grid values in \code{XY}.  Defaults to grid-based approximation using \code{XY}.}
}
\value{
An \code{xDens} object.  See \code{\link{xDens}}.
}
\description{
Constructs an \code{xDens} object using a pre-specified matrix of grid and density values.
}
\examples{
# simulate data from Cauchy(0, 1) distribution
nsamples = 1e4
X = seq(from=-5, to=4, by=0.1) # grid
Y = dcauchy(X)
xDens = matrixXD(cbind(X, Y))
par(mfrow = c(1,2), mar = c(4,4,2,.5)+.1)

# pdf and sampling check
Xsim <- rXD(nsamples, xDens = xDens) # xDensity: random sampling
hist(Xsim, breaks = 100, freq = FALSE,
     xlab = "x", main = "PDF")
curve(dXD(x, xDens = xDens), add = TRUE, col = "red") # xDensity PDF
curve(dcauchy(x), add = TRUE, col = "blue") # true PDF
abline(v = xDens$xrng, lty = 2) # grid endpoints
legend("topleft", c("rXD", "dXD", "Cauchy(0,1)", "xRange"),
       pch = c(22,22,22,NA), pt.cex = 1,
       pt.bg = c("white", "red", "blue", "black"),
       lty = c(NA, NA, NA, 2))

# cdf check
curve(pXD(x, xDens), # xDens CDF
      from = min(Xsim), to = max(Xsim), col = "red",
      xlab = "x", main = "CDF", ylab = "Cumulative Probability")
curve(pcauchy(x), add = TRUE, col = "blue") # true CDF
abline(v = xDens$xrng, lty = 2) # grid endpoints
legend("bottomright", c("pXD", "Cauchy(0,1)", "xRange"),
       pch = c(22,22,NA), pt.cex = 1.5,
       pt.bg = c("red", "blue", "black"),
       lty = c(NA, NA, 2))
}
