% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xDensity.R
\name{normalXD}
\alias{normalXD}
\title{Normal Extended Density.}
\usage{
normalXD(x, n = 512, from, to, mean, sd)
}
\arguments{
\item{x}{vector of samples to which to fit the normal distribution.}

\item{n, from, to}{optional arguments to set up the grid on which to store the density estimate.}

\item{mean, sd}{optional mean and standard deviation arguments for the extended density.  Default to the mean and standard deviation of \code{x}.}
}
\value{
an \code{xDens} object.
}
\description{
Normal Extended Density.
}
