% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powFit.R
\name{powFit}
\alias{powFit}
\title{Maximum likelihood estimate of the generalized Box-Cox transform.}
\usage{
powFit(x, alpha = NA, interval = c(-5, 5), ...)
}
\arguments{
\item{x}{Vector of samples from density.}

\item{alpha}{Optional value of the offset parameter.  \code{alpha = FALSE} sets \code{alpha = 1 - min(x)}, thereby guaranteeing that \code{z = x + alpha >= 1}.  This or any scalar value of \code{alpha} optimizes as a function of \code{lambda} only.  \code{alpha = NA} jointly optimizes for \code{lambda} and \code{alpha}.}

\item{interval}{Range of \code{lambda} values for one dimensional optimization.}

\item{...}{Additional arguments to pass to \code{optimize} or \code{optim}, for 1- or 2-parameter optimization.}
}
\value{
MLEs for \code{lambda} and possibly \code{alpha} as well.
}
\description{
Maximum likelihood estimate of the generalized Box-Cox transform.
}
\details{
The likelihood for optimization is
\preformatted{L(lambda, alpha | x) = prod(dnorm(z(x | lambda, alpha)) *  |dz(x | lambda, alpha) / dx|)},
where \code{z(x | lambda, alpha)} is the Box-Cox transformation.
}
