% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powFit.R
\name{powTrans}
\alias{powTrans}
\title{Generalized Power (or Box-Cox) Transformation.}
\usage{
powTrans(x, lambda = 0, alpha = 0, normalize = FALSE, jacobian = FALSE,
  debug = FALSE)
}
\arguments{
\item{x}{Vector of values at which to compute the transformation.}

\item{lambda}{Exponent of the transformation.  See details.}

\item{alpha}{Offset of the transformation.  See details.}

\item{normalize}{Logical; if TRUE divides by the geometric mean.  See details.}

\item{jacobian}{Logical; if TRUE calculates the Jacobian \code{|dz / dx|}, which converts transformed density values back to the original scale.}
}
\value{
The vector \code{z} of transformed values, and optionally the Jacobian of the inverse transformation.  See details.
}
\description{
Generalized Power (or Box-Cox) Transformation.
}
\details{
The Generalized Power or Box-Cox transformation is
\deqn{\code{z = ((x + alpha)^lambda - 1) / (lambda * C^(lambda-1))}, \code{lambda != 0},}
\deqn{\code{z = C * log(x + alpha)}, \code{lambda == 0},}
where \code{C} is the Geometric mean:
\deqn{\code{C = exp(mean(log(x + alpha)))}.}
Note that \code{C} is only calculated if \code{normalize = TRUE}.
}
