% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgc4.R
\name{dgc4}
\alias{dgc4}
\title{Univariate 4th-order Gram-Charlier density approximation.}
\usage{
dgc4(x, cmom, rm.neg = TRUE, log = FALSE)
}
\arguments{
\item{x}{Vector of density quantiles.}

\item{cmom}{Vector of first 4 central moments of distribution.}

\item{rm.neg}{Logical; if \code{TRUE} clips the density at smallest calculated non-negative value.}

\item{log}{Logical; if \code{TRUE} returns approximating density on log scale.}
}
\value{
Vector of density evaluations.
}
\description{
Univariate 4th-order Gram-Charlier density approximation.
}
\examples{
df <- 3
nsamples <- 100
X <- rchisq(nsamples, df=df) # iid samples from the Chi-Sq(df)

# calculate central moments
cmom <- mean(X)
cmom <- c(cmom, mean((X-cmom)^2), mean((X-cmom)^3), mean((X-cmom)^4))

curve(dgc4(x, cmom = cmom, rm.neg = TRUE), from = 0, to = 10)
}
