% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussCop-package.R
\docType{package}
\name{GaussCop}
\alias{GaussCop}
\alias{GaussCop-package}
\title{Simulation, Inference, and Density Evaluation for the Gaussian Copula Distribution.}
\description{
The GaussCop package proposes a simple, unified framework to provide the basic 
d/p/q/r functions for arbitrary continuous one-dimensional distributions. This can be useful 
for working with various Copula models, including the Gaussian Copula, for which specific 
functions are provided.
}
