% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc4.R
\name{gc4XD}
\alias{gc4XD}
\title{Combined Box-Cox Transformation + Gram-Charlier Density Approximation.}
\usage{
gc4XD(x, lambda = NULL, alpha = 0, cmom = NULL, trim = 0.01, n = 512,
  from, to, mean, sd, ..., debug = FALSE)
}
\arguments{
\item{x}{Sample from density to approximate.}

\item{lambda}{Exponent of Box-Cox transform.  If NULL it is estimated from \code{x}.  See details.}

\item{alpha}{Offset of the Box-Cox transform.  Default is no offset.  See details.}

\item{cmom}{Optional vector of first 4 central moments.  If NULL these are estimated from \code{x}.}

\item{trim}{Scalar between 0 and 1; removes the \code{trim} fraction of extreme values from \code{x} for estimation of \code{lambda} and \code{cmom}, which are very sensitive to outliers.  \code{trim = FALSE} does not trim any values.}

\item{n, from, to}{specifies a grid of values on which to evaluate the density.}

\item{mean, sd}{mean and standard deviation for extended density.}

\item{...}{additional parameters to Box-Cox fitting function \code{\link{powFit}}.}
}
\value{
An \code{xDens} object.
}
\description{
Combined Box-Cox Transformation + Gram-Charlier Density Approximation.
}
\details{
\code{x} is first standardized to \code{z = x/sd(x) - min(x/sd(x), from) + 1}, before the Box-Cox transform is applied.

\code{lambda} can be estimated from the data, but for stability \code{alpha} must be provided.
}

