% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc4XD.R
\name{gc4XD}
\alias{gc4XD}
\title{Construct an \code{xDensity} representation of a combined Box-Cox/Gram-Charlier density approximation.}
\usage{
gc4XD(x, lambda = NULL, alpha = 0, cmom = NULL, trim = 0.01,
  n = 512, from, to, mean, sd, ...)
}
\arguments{
\item{x}{Vector of random samples from density to approximate.}

\item{lambda}{Exponent of Box-Cox transform.  If \code{NULL} it is estimated from \code{x}.  See Details.}

\item{alpha}{Offset of the Box-Cox transform.  Default is no offset.  See Details.}

\item{cmom}{Optional vector of first 4 central moments of \code{x}.  If \code{NULL} these are estimated from \code{x}.}

\item{trim}{Scalar between 0 and 1; removes the \code{trim} fraction of extreme values from \code{x} for estimation of \code{lambda} and \code{cmom}, which are very sensitive to outliers.  \code{trim = FALSE} does not trim any values.}

\item{n, from, to}{Specifies a grid of values on which to evaluate the density (see \code{\link[stats]{density}}).}

\item{mean, sd}{Optional mean and standard deviation for \code{xDensity} representation.}

\item{...}{Additional parameters to Box-Cox fitting function \code{\link{powFit}}.}
}
\value{
An \code{xDensity} object.
}
\description{
Construct an \code{xDensity} representation of a combined Box-Cox/Gram-Charlier density approximation.
}
\details{
\code{x} is first standardized to \code{z = x/sd(x) - min(x/sd(x), from) + 1}, before the Box-Cox transform is applied.

For details on the Box-Cox transformation and Gram-Charlier approximation, see \code{\link{powFit}} and \code{\link{dgc4}} respectively.
}
\examples{
# xDensity approximation to a noncentral-t distribution

# true parameters
lambda <- rnorm(1) # noncentrality parameter
nu <- runif(1, 4, 6) # degrees of freedom

# simulate data (note the small sample size)
x <- rt(500, df = nu, ncp = lambda)

# xDensity approximation
xDensK <- kernelXD(x) # kernel smoothing
xDensG <- gc4XD(x) # gc4 approximation

# true vs approximate PDFs
xlim <- qt(c(.005, .995), df = nu, ncp = lambda) # range for plot
curve(dt(x, df = nu, ncp = lambda),
      from = xlim[1], to = xlim[2], ylab = "Density")
curve(dXD(x, xDens = xDensK), add = TRUE, col = "red")
curve(dXD(x, xDens = xDensG), add = TRUE, col = "blue")
legend("topleft", legend = c("True PDF", "xDensity: kernel", "xDensity: gc4"),
       fill = c("black", "red", "blue"))
}
\seealso{
\code{\link{powFit}}, \code{\link{dgc4}}, \code{\link{xDensity}}.
}
