% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcop.R
\name{gcop}
\alias{gcop}
\alias{dgcop}
\alias{rgcop}
\title{The Gaussian Copula Distribution.}
\usage{
dgcop(X, gCop, log = FALSE, decomp = FALSE)

rgcop(n, gCop)
}
\arguments{
\item{X}{\code{n x p} values at which to evaluate the \code{p}-dimensional density.}

\item{gCop}{the \code{gcop} model specification.}

\item{decomp}{logical; if TRUE returns the normalized residuals \code{Z}, their log-density \code{zlpdf}, and the log-jacobian \code{zljac}, such that the total log-density is \code{xldens = zldens + zljac}.  See details.}

\item{n}{number of random samples to draw.}
}
\value{
A vector of densities or matrix of random samples.
}
\description{
The Gaussian Copula Distribution.
}
\details{
The density of Gaussian Copula distribution is
\deqn{
g(x) = \frac{\psi(z \mid \Rho) \prod_{i=1}^d f_i(x_i)}{\prod_{i=1}^d\phi(z_i)},
}{
g(x) = \psi(z | R) \prod_{i=1}^d f_i(x_i)/\phi(z_i),
}
\deqn{
z_i = \Phi^{-1}(F_i(x_i)),
}{
z_i = \Phi^(-1)(F_i(x_i)),
}
where \eqn{\psi(z \mid \Rho)}{\psi(z | R)} is the PDF of a multivariate normal with mean 0 and variance \eqn{\Rho}{R}, \eqn{f_i(x_i)} and \eqn{F_i(x_i)} are the marginal PDF and CDF of variable \eqn{i}, and \eqn{\phi(z)} and \eqn{\Phi(z)} are the PDF and CDF of a standard normal.
}
