% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcopFit.R
\name{gcopFit}
\alias{gcopFit}
\title{Fit a Gaussian Copula Model.}
\usage{
gcopFit(X, Rho, fitXD = c("kernel", "gc4", "normal"), ...)
}
\arguments{
\item{X}{either an \code{N x p} data matrix or a \code{p}-length list of \code{XDens} objects.}

\item{Rho}{\code{p x p} correlation matrix (i.e., 1's on the diagonal).  Optional if \code{X} is a matrix.}

\item{fitXD}{string specifying method to fit marginals (see details).}

\item{...}{additional arguments to pass to the methods of \code{fitXD} (see details).}
}
\value{
An object of class \code{gcop} (see details).
}
\description{
Fit a Gaussian Copula Model.
}
\details{
A \code{gcop} object is a list with elements:
\itemize{
  \item \code{XDens}: a list of \code{xDens} objects specifying each marginal distribution (see \code{\link{xDens}}).
  \item \code{Rho}: the correlation matrix between the normalized quantiles.
}
}

