% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xDensity.R
\name{kernelXD}
\alias{kernelXD}
\title{Extended Kernel Density Approximation}
\usage{
kernelXD(x, n = 512, from, to, mean, sd, any0 = FALSE, ...)
}
\arguments{
\item{x}{vector of samples from the underlying distribution.}

\item{n, from, to}{optional arguments to \code{\link{stat::density}} which are used to set the grid on which to evaluate the kernel estimator.}

\item{mean, sd}{optional mean and standard deviation arguments for the extended density.  Default to the mean and standard deviation of \code{x}.}

\item{any0}{logical; if FALSE forces the support of the density to be the real line.}

\item{...}{additional arguments to \code{\link{stat::density}}.}
}
\value{
an \code{xDens} object.
}
\description{
Extended Kernel Density Approximation
}
