% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgcopCond.R
\name{rgcopCond}
\alias{rgcopCond}
\title{Simulation of a Gaussian Copula with Conditioning.}
\usage{
rgcopCond(n, gCop, XCond, iCond, debug = FALSE)
}
\arguments{
\item{n}{Integer number of random draws}

\item{gCop}{An object of class \code{gcop} describing the joint Gaussian Copula distribution on all variables.}

\item{XCond}{Vector or matrix; values of the variables on which to condition.}

\item{iCond}{Vector of integers, logicals, or characters specifying which are the random variables on which to condition.}
}
\value{
An \code{n x (nrv-nCond)} matrix containing the draws from the conditional distribution.
}
\description{
Simulation of a Gaussian Copula with Conditioning.
}

