% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xDens.R
\name{xDens}
\alias{xDens}
\alias{dXD}
\alias{pXD}
\alias{qXD}
\alias{rXD}
\title{Extended-Density Objects.}
\usage{
dXD(x, xDens, log = FALSE, debug = FALSE)

pXD(q, xDens, lower.tail = TRUE, log.p = FALSE)

qXD(p, xDens, lower.tail = TRUE, log.p = FALSE, debug = FALSE)

rXD(n, xDens)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{xDens}{object containing the extended-density (\code{xDens}) representation.}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X <= x]} otherwise, \code{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Extended-Density Objects.
}
\details{
An \code{xDens} object is a list with the following elements:
\itemize{
  \item \code{xrng}, \code{ndens}: range and number of gridpoints defining the main density region, i.e. \code{xseq = seq(xrng[1], xrng[2], len = xn)}.
  \item \code{ypdf}, \code{ylpdf}, \code{ycdf}: density, log-density, and cdf on the grid.
  \item \code{mean}, \code{sd}: mean and standard deviation of a normal to use outside the density range.
}
}
